# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_coindistro_backend_internal_identity_models.ActivityLogResponse:
    properties:
      action:
        type: string
      created_at:
        type: string
      details:
        additionalProperties: true
        type: object
      device_id:
        type: string
      id:
        type: string
      ip_address:
        type: string
    type: object
  github_com_coindistro_backend_internal_identity_models.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.UserResponse'
    type: object
  github_com_coindistro_backend_internal_identity_models.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  github_com_coindistro_backend_internal_identity_models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_coindistro_backend_internal_identity_models.PasswordResetComplete:
    properties:
      password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
  github_com_coindistro_backend_internal_identity_models.PasswordResetRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  github_com_coindistro_backend_internal_identity_models.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  github_com_coindistro_backend_internal_response.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  github_com_coindistro_backend_internal_response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIError'
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_coindistro_backend_internal_response.Meta'
      success:
        type: boolean
    type: object
...
host: coindistro.onrender.com
info:
  contact:
    email: support@coindistro.com
    name: Coindistro Support
    url: https://coindistro.com/support
  description: Coindistro — One Platform. Everything Crypto. API backend for Africa's
    next-generation crypto financial ecosystem.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://coindistro.com/terms
  title: Coindistro API
  version: 1.0.0
paths:
  /activity:
    get:
      description: Get security and activity history for the current user
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.ActivityLogResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get activity log
      tags:
      - Security
  /auth/change-password:
    put:
      consumes:
      - application/json
      description: Change your current password
      parameters:
      - description: Current and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Security
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send a password reset link to your email
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      summary: Request password reset
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.AuthResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      summary: Login to your account
      tags:
      - Authentication
  /auth/logout:
    post:
      description: Terminate the current session
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.AuthResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      summary: Refresh access token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register with referral code. Invite-only when enabled.
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      summary: Register a new account
      tags:
      - Authentication
  /auth/resend-verification:
    post:
      description: Resend the email verification link
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Resend verification email
      tags:
      - Authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Complete password reset with token and new password
      parameters:
      - description: Reset token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.PasswordResetComplete'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      summary: Reset password
      tags:
      - Authentication
  /auth/verify-email:
    get:
      description: Verify email using the token sent to your email
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      summary: Verify email address
      tags:
      - Authentication
  /devices:
    get:
      description: Get all trusted devices for the current user
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.DeviceResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List trusted devices
      tags:
      - Devices
  /devices/{id}:
    delete:
      description: Remove a trusted device by ID
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Remove a device
      tags:
      - Devices
...
  /invitations:
    get:
      description: Get all invitations sent by the current user
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.InvitationResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List invitations
      tags:
      - Invitations
    post:
      consumes:
      - application/json
      description: Send an invitation email. Consumes one invitation credit.
      parameters:
      - description: Invitation details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.SendInvitationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.InvitationResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Send invitation
      tags:
      - Invitations
  /referrals/dashboard:
    get:
      description: Get referral code, stats, and tree
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.ReferralDashboardResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get referral dashboard
      tags:
      - Referrals
  /sessions:
    get:
      description: Get all active sessions for the current user
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.SessionResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List active sessions
      tags:
      - Sessions
  /sessions/{id}:
    delete:
      description: Terminate a specific session by ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Terminate a session
      tags:
      - Sessions
  /sessions/terminate-all:
    post:
      description: Terminate all sessions except the current one
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Terminate all sessions
      tags:
      - Sessions
  /users/check-email:
    get:
      description: Check if an email is already registered
      parameters:
      - description: Email to check
        in: query
        name: email
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      summary: Check email availability
      tags:
      - Users
  /users/check-username:
    get:
      description: Check if a username is already taken
      parameters:
      - description: Username to check
        in: query
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
      summary: Check username availability
      tags:
      - Users
  /users/me:
    get:
      description: Returns the authenticated user's profile
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.UserResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update profile fields like display_name, username, country, timezone
      parameters:
      - description: Profile fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_coindistro_backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_coindistro_backend_internal_identity_models.UserResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - Users
schemes:
- https
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"